book-service/requirements.txt
Django==5.1.1
djangorestframework==3.14.0
mysqlclient==2.2.0
gunicorn==21.2.0
django-redis==5.4.0
kafka-python-ng==2.2.2
